% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwd1asR0.R
\name{bwd1asR0}
\alias{bwd1asR0}
\title{bwd1asR0}
\usage{
bwd1asR0(R, dim, M)
}
\arguments{
\item{R}{The correlation matrix of \eqn{\mathbf{X}}.}

\item{dim}{The vector of dimensions \eqn{(d_{1},...,d_{k})}, in ascending order.}

\item{M}{The sample size.}
}
\value{
A sample of size M from the asymptotic distribution of the first Bures-Wasserstein dependence coefficient \eqn{\mathcal{D}_{1}} under independence of \eqn{\mathbf{X}_{1},...,\mathbf{X}_{k}}.
}
\description{
Given a q-dimensional random vector \eqn{\mathbf{X} = (\mathbf{X}_{1},...,\mathbf{X}_{k})} with \eqn{\mathbf{X}_{i}} a \eqn{d_{i}} dimensional random vector, i.e., \eqn{q = d_{1} + ... + d_{k}},
this function simulates a sample from the asymptotic distribution of the correlation-based Bures-Wasserstein coefficient \eqn{\mathcal{D}_{1}}
between \eqn{\mathbf{X}_{1},...,\mathbf{X}_{k}} given the entire correlation matrix \eqn{\mathbf{R}} is equal to \eqn{\mathbf{R}_{0}} (correlation matrix under independence of \eqn{\mathbf{X}_{1},...,\mathbf{X}_{k}}).
The argument dim should be in ascending order.
This function requires importation of the python modules "numpy" and "scipy".
}
\examples{
q = 5
dim = c(2,3)
R = 0.5^(abs(matrix(1:q-1,nrow = q, ncol = q, byrow = TRUE) - (1:q-1))) # AR(1) correlation matrix with correlation 0.5
R0 = createR0(R,dim)
sample = bwd1asR0(R0,dim,1000)
}
