% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvomega.R
\name{cvomega}
\alias{cvomega}
\title{cvomega}
\usage{
cvomega(sample, omegas, K)
}
\arguments{
\item{sample}{A sample from a q-dimensional random vector \eqn{\mathbf{X}} (n x q matrix with observations in rows, variables in columns).}

\item{omegas}{A grid of candidate penalty parameters in \eqn{[0,1]}.}

\item{K}{The number of folds to be used.}
}
\value{
The optimal ridge penalty parameter minimizing the cross-validation error.
}
\description{
This functions selects the omega tuning parameter for ridge penalization of the empirical Gaussian copula correlation matrix via cross-validation.
             The objective function is the Gaussian log-likelihood, and a grid search is performed using K folds.
}
\examples{
q = 10
n = 50
R = 0.5^(abs(matrix(1:q-1,nrow = q, ncol = q, byrow = TRUE) - (1:q-1))) # AR(1) correlation matrix with correlation 0.5
sample = mvtnorm::rmvnorm(n,rep(0,q),R,method = "chol") # Sample from multivariate normal distribution
omega = cvomega(sample = sample,omegas = seq(0.01,0.999,len = 50),K = 5) # 5-fold cross-validation with Gaussian likelihood as loss for selecting omega
R_est = estR(sample,omega = omega)
}
