% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tensorflow.R
\name{install_tensorflow}
\alias{install_tensorflow}
\title{install_tensorflow}
\usage{
install_tensorflow(..., envname = "r-tensorflow")
}
\description{
This function installes a python virtual environment.
}
