% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miStudent.R
\name{miStudent}
\alias{miStudent}
\title{miStudent}
\usage{
miStudent(R, dim, nu)
}
\arguments{
\item{R}{The correlation matrix of \eqn{\mathbf{X}}.}

\item{dim}{The vector of dimensions \eqn{(d_{1},...,d_{k})}.}

\item{nu}{The degrees of freedom.}
}
\value{
The Student-t mutual information between \eqn{\mathbf{X}_{1},...,\mathbf{X}_{k}}.
}
\description{
Given a q-dimensional random vector \eqn{\mathbf{X} = (\mathbf{X}_{1},...,\mathbf{X}_{k})} with \eqn{\mathbf{X}_i} a \eqn{d_{i}} dimensional random vector, i.e., \eqn{q = d_{1} + ... + d_{k}},
this function computes the Student-t mutual information between \eqn{\mathbf{X}_{1},...,\mathbf{X}_{k}} given the entire correlation matrix \eqn{\mathbf{R}} and the degrees of freedom nu.
}
\examples{
q = 10
dim = c(1,2,3,4)
R = 0.5^(abs(matrix(1:q-1,nrow = q, ncol = q, byrow = TRUE) - (1:q-1))) # AR(1) correlation matrix with correlation 0.5
nu = 7
miStudent(R,dim,nu)
}
