% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tensorflow.R
\name{install_tensorflow}
\alias{install_tensorflow}
\title{install_tensorflow}
\usage{
install_tensorflow(envname = "r-tensorflow")
}
\arguments{
\item{envname}{Name of the environment.}
}
\value{
No return value, used for creating a python virtual environment.
}
\description{
This function installs a python virtual environment.
}
